Shared tests between platforms
