+++ example/package.json
@@ -12,8 +12,8 @@
     "build:android": "npm-run-all clean build:android:*",
     "build:ios": "cd ios && xcodebuild build -workspace example.xcworkspace -scheme example -sdk iphonesimulator ONLY_ACTIVE_ARCH=NO -destination 'platform=iOS Simulator,name=iPhone 6' -configuration Release -derivedDataPath build",
     "build": "npm-run-all build:*",
-    "test:android": "PLATFORM_NAME=android APP_PATH=`pwd`/android/app/build/outputs/apk/app-release.apk TESTS_PATH=test/*_test_*.android.js babel-node ./test/utils/run.js",
-    "test:ios": "PLATFORM_NAME=ios NO_RESET=true AUTOMATION_NAME=xcuitest APP_PATH=`pwd`/ios/build/Build/Products/Release-iphonesimulator/example.app TESTS_PATH=test/*_test_*.ios.js babel-node ./test/utils/run.js",
+    "test:android": "PLATFORM_NAME=android APP_PATH=`pwd`/android/app/build/outputs/apk/app-release.apk TESTS_PATH=test/*_test_*.js IGNORE_PATH=test/*_test_*.ios.js babel-node ./test/utils/run.js",
+    "test:ios": "PLATFORM_NAME=ios NO_RESET=true AUTOMATION_NAME=xcuitest APP_PATH=`pwd`/ios/build/Build/Products/Release-iphonesimulator/example.app TESTS_PATH=test/*_test_*.js IGNORE_PATH=test/*_test_*.android.js babel-node ./test/utils/run.js",
     "test": "npm-run-all test:*"
   },
   "dependencies": {
@@ -37,6 +37,7 @@
     "imgur": "^0.1.7",
     "node-simctl": "^3.4.4",
     "npm-run-all": "^3.1.1",
+    "require-dir-all": "^0.4.10",
     "tap-diff": "^0.1.1",
     "tape": "^4.6.2",
     "tape-async": "^2.1.1",
