--- example/test/03_test_custom_card.android.js
@@ -1,32 +1,0 @@
-import test from 'tape-async'
-import helper from './utils/helper'
-
-const { driver, idFromAccessId } = helper
-
-test('Test if user can use Custom Card params', async(t) => {
-  const cardFormTabId = idFromAccessId('headerTab_2')
-  const cardFormButtonId = idFromAccessId('customCardButton')
-  const tokenId = idFromAccessId('customCardToken')
-
-  try {
-    await driver.waitForVisible(cardFormTabId, 70000)
-    await driver.click(cardFormTabId)
-
-    await driver.waitForVisible(cardFormButtonId, 10000)
-
-    t.pass('User should see `Pay with custom params` button')
-
-    await driver.click(cardFormButtonId)
-
-    t.pass('User should be able to tap on `Pay with custom params` button')
-
-    await driver.waitForVisible(tokenId, 500000)
-
-    t.pass('User should see token')
-  } catch (error) {
-    await helper.screenshot()
-    await helper.source()
-
-    throw error
-  }
-})
